+++ lib/zaikio/mission_control/finishings/base.rb
@@ -1,0 +1,8 @@
+module Zaikio
+  module MissionControl
+    module Finishings
+      class Base < Zaikio::MissionControl::Base
+      end
+    end
+  end
+end