+++ lib/zaikio/mission_control/finishings/foil_stamp.rb
@@ -1,0 +1,8 @@
+module Zaikio
+  module MissionControl
+    module Finishings
+      class FoilStamp < Base
+      end
+    end
+  end
+end