+++ lib/zaikio/mission_control/finishings/glue.rb
@@ -1,0 +1,8 @@
+module Zaikio
+  module MissionControl
+    module Finishings
+      class Glue < Base
+      end
+    end
+  end
+end