+++ lib/zaikio/mission_control/finishings/hole.rb
@@ -1,0 +1,8 @@
+module Zaikio
+  module MissionControl
+    module Finishings
+      class Hole < Base
+      end
+    end
+  end
+end