+++ lib/zaikio/mission_control/finishings/lamination.rb
@@ -1,0 +1,8 @@
+module Zaikio
+  module MissionControl
+    module Finishings
+      class Lamination < Base
+      end
+    end
+  end
+end