+++ lib/zaikio/mission_control/finishings/perforation.rb
@@ -1,0 +1,8 @@
+module Zaikio
+  module MissionControl
+    module Finishings
+      class Perforation < Base
+      end
+    end
+  end
+end