+++ lib/zaikio/mission_control/finishings/saddle_stitch.rb
@@ -1,0 +1,8 @@
+module Zaikio
+  module MissionControl
+    module Finishings
+      class SaddleStitch < Base
+      end
+    end
+  end
+end