+++ lib/zaikio/mission_control/finishings/strip_binding.rb
@@ -1,0 +1,8 @@
+module Zaikio
+  module MissionControl
+    module Finishings
+      class StripBinding < Base
+      end
+    end
+  end
+end