+++ lib/zaikio/mission_control/finishings/thread_sewing.rb
@@ -1,0 +1,8 @@
+module Zaikio
+  module MissionControl
+    module Finishings
+      class ThreadSewing < Base
+      end
+    end
+  end
+end